CLI: fix linter errors
